Rename : caffeine 폴더 생성
caffeine 폴더 생성해 관련 config 및 CacheType 클래스 이동
